feat. exceptions

+++ src/main/java/br/com/helper/oop/contabancaria/ContaBancaria.java
@@ -1,26 +1,28 @@
 package br.com.helper.oop.contabancaria;
 
+import java.math.BigDecimal;
+
 public class ContaBancaria {
 
-	private int saldo;
+	private double saldo;
 
-	public ContaBancaria(int saldoInicial) {
-	        this.saldo = saldoInicial;
+	public ContaBancaria(BigDecimal saldoInicial) {
+	        this.saldo = saldoInicial.doubleValue();
 	    }
 
-	public void depositar(int valor) {
-		if (valor > 0) {
-			saldo += valor;
+	public void depositar(BigDecimal valor) {
+		if (valor.floatValue() > 0) {
+			saldo += valor.doubleValue();
 		}
 	}
 
-	public void sacar(int valor) {
-		if (valor > 0 && valor <= saldo) {
-			saldo -= valor;
+	public void sacar(BigDecimal valor) {
+		if (valor.doubleValue() > 0 && valor.doubleValue() <= saldo) {
+			saldo -= valor.doubleValue();
 		}
 	}
 
-	public int getSaldo() {
+	public double getSaldo() {
 		return saldo;
 	}
 }
